requirements.txt
aioredis==1.3.1
amqp==5.1.1
arabic-reshaper==2.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
autopep8==1.6.0
billiard==3.6.4.0
branca==0.4.2
celery==5.2.7
certifi==2020.12.5
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
crispy-bootstrap5==0.6
cryptography==36.0.0
cx-Oracle==8.2.0
cycler==0.11.0
Cython==0.29.33
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.4
django-authtools==1.7.0
django-betterforms==1.2
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-embed-video==1.4.0
django-extensions==3.1.5
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.8.0
django-mathfilters==1.0.0
django-model-utils==4.2.0
django-multiselectfield==0.1.12
django-phonenumber-field==7.0.2
django-widget-tweaks==1.4.8
django3-livesync==1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
drf-extensions==0.7.1
environ==1.0
environs==9.3.5
et-xmlfile==1.1.0
filelock==3.6.0
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
folium==0.12.1
future==0.18.2
GDAL @ file:///C:/Users/zheny/downloads/GDAL-3.4.3-cp39-cp39-win_amd64.whl
geographiclib==1.50
geopy==2.1.0
graphviz==0.20
greenlet==1.1.2
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.0
kiwisolver==1.3.2
kombu==5.2.4
MarkupPy==1.14
MarkupSafe==2.0.0
marshmallow==3.14.1
matplotlib==3.4.3
msgpack==1.0.3
ndg-httpsclient==0.5.1
networkx==2.7.1
numpy==1.20.3
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.3.4
phonenumbers==8.13.6
Pillow==8.2.0
pipreqs==0.4.11
platformdirs==2.5.1
prompt-toolkit==3.0.31
psycopg2==2.9.3
pusher==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydot==1.4.2
pydot-ng==2.0.0
pydub==0.25.1
pyit2fls==0.7.9
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
PyPDF2==1.26.0
pyswip==0.2.11
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-utils==3.5.2
pytz==2022.5
PyYAML==6.0
redis==4.0.2
reportlab==3.6.1
requests==2.25.1
scikit-fuzzy==0.4.2
scipy==1.7.2
seaborn==0.11.2
selenium==3.141.0
service-identity==21.1.0
six==1.16.0
SpeechRecognition==3.8.1
SQLAlchemy==1.4.27
sqlparse==0.4.1
stop-words==2018.7.23
tablib==3.2.1
tabulate==0.8.9
telebot==0.0.4
toml==0.10.2
tornado==6.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==4.0.0
tzdata==2022.1
urllib3==1.26.4
validate-email==1.3
vine==5.0.0
virtualenv==20.13.3
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
whitenoise==5.3.0
wordcloud==1.8.1
wrapt==1.13.3
WTForms==3.0.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
zope.interface==5.4.0
